Rest Client refactored
